// src/app/models/financial-management/financeReceiptConfigData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface FinanceReceiptConfigData {
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    /** ชื่อผู้มีอำนาจลงนาม (ยังไม่ใช้งาน) */
    approvedName?: string;
    /** ตำแหน่งผู้มีอำนาจลงนาม (ยังไม่ใช้งาน) */
    approvedPosition?: string;
    /** รูป Signature ผู้มีอำนาจลงนาม (ยังไม่ใช้งาน) */
    approvedSignaturePath?: string;
    /** ที่อยู่หน่วยงาน (อังกฤษ) */
    depAddressEn?: string;
    /** ที่อยู่หน่วยงาน (ไทย) */
    depAddressTh?: string;
    /** ชื่อหน่วยงาน (อังกฤษ) */
    depNameEn?: string;
    /** ชื่อหน่วยงาน (ไทย) */
    depNameTh?: string;
    /** เลขประจำตัวผู้เสียภาษีของหน่วยงาน */
    depTaxId?: string;
    first?: number;
    /** path logo ใบเสร็จรับเงิน */
    logoPath?: string;
    mode?: string;
    /** PK */
    receiptConfigId?: number;
    /** คำลงท้ายใบเสร็จ (อังกฤษ) */
    receiptNoteEn?: string;
    /** คำลงท้ายใบเสร็จ (ไทย) */
    receiptNoteTh?: string;
    /** prefix เลขใบเสร็จรับเงิน */
    receiptPrefix?: string;
    /** หมายเหตุใบเสร็จ */
    receiptRemark?: string;
    rowNum?: number;
    size?: number;
    /** ชื่อผู้รับเงิน */
    staffName?: string;
    /** ตำแหน่งผู้ชำระเงิน */
    staffPosition?: string;
    /** รูป Signature ผู้รับเงิน */
    staffSignaturePath?: string;

    filename?: string;
    prefix?: string;
	module?: number;
    fullpath?: string;

    staffSignaturePrefix?: string;
    staffSignatureModule?: number;
    logoPrefix?: string;
    logoModule?: number;

}